Lukkeliste stability and performance (#59)
* FIX: Lukkeliste, adjust matr-logic

we now only query once to get the relevant matr from database using multipoint, instead of multiple calls bogging down the instance.

added wkb as a parameter for datahub.js that mimics DAWA output, but accepts whaterver-gemetry-fits-in-ewkb-format geometry.

minor typos and formatting.

dont cache results from serviceworker

* Remove Maxfeatures from server implementation

we get the information as a massive multipoint, and handle the getting of information from frontend. the number of features are a product of the analysis, so we can selectively choose to ignore it.

* Async some more + TooManyFeatures

implement maxfeatures, to keep system from automatically getting all adresser a once.

* refactor TooManyFeatures into state

* progress + less insanity

* throw errors, optimize getting stuff from sql

* add maxparcel limit

* better error messages, need post endpoint for wkb

* themes, limits and more. post endpoint for datahub
