+++ extensions/blueidea/browser/i8n.js
@@ -1,0 +1,124 @@
+var dict = {
+    Info: {
+      da_DK:
+        "BlueIdea / Lukkeliste er et modul der hjælper med at finde relevante adresser til en lukning. Du kan tegne et område på kortet, og få en liste over adresser i området. Du kan også udpege et punkt på ledningsnettet, og få en liste af adresser der er forbundet til det punkt. Listerne med adresser kan derefter bruges til at sende beskeder ud med BlueIdea (Kræver aftale med BlueIdea).",
+      en_US:
+        "BlueIdea / Valve list is a module that helps to find relevant addressess for a valve closure. You can draw an area on the map, and get a list of addresses in the area. You can also select a point on the network, and get a list of addresses connected to that point. The lists of addresses can then be used to send messages with BlueIdea (Requires agreement with BlueIdea).",
+    },
+    "Plugin Tooltip": {
+      da_DK: "BlueIdea / Lukkeliste",
+      en_US: "BlueIdea / Valve list",
+    },
+    MissingLogin: {
+      da_DK:
+        "NB: Du skal være logget ind for at kunne bruge funktionen, og din konfiguration skal indeholde et brugerid.",
+      en_US:
+        "Please log in and set a user id in the configuration file to use this function",
+    },
+    Login: {
+      da_DK: "Log ind",
+      en_US: "Log in",
+    },
+    "Select point on map": {
+      da_DK: "Udpeg punkt på ledningsnet",
+      en_US: "Select point on map",
+    },
+    "Select point for alarmkabel": {
+      da_DK: "Beregn alarm",
+      en_US: "Calculate alarm",
+    },
+    "Found parcels": {
+      da_DK: "Fundne matrikler i område",
+      en_US: "Found parcels in area",
+    },
+    "Found addresses": {
+      da_DK: "Alle adresser fundet",
+      en_US: "All addresses found",
+    },
+    "Error in search": {
+      da_DK: "Fejl i søgning af matrikler",
+      en_US: "Error in search for parcels",
+    },
+    "Draw area": {
+      da_DK: "Tegn områder",
+      en_US: "Draw area",
+    },
+    "Select area": {
+      da_DK: "Udpeg",
+      en_US: "Select",
+    },
+    "Show results": {
+      da_DK: "Resultater",
+      en_US: "Results",
+    },
+    "Waiting to start": {
+      da_DK: "Henter jordstykker",
+      en_US: "Fetching parcels",
+    },
+    "Go to blueidea": {
+      da_DK: "Opret, og gå til blueidea",
+      en_US: "Create and go to blueidea",
+    },
+    "Valve list": {
+      da_DK: "Lukkeliste",
+      en_US: "Valve list",
+    },
+    "Download addresses": {
+      da_DK: "Download adresser",
+      en_US: "Download addresses",
+    },
+    "Get addresses": {
+      da_DK: "Hent adresser",
+      en_US: "Get addresses",
+    },
+    "Download valves": {
+      da_DK: "Download ventil-liste",
+      en_US: "Download valve list",
+    },
+    NotAllowedBlueIdea: {
+      da_DK: "Du har ikke adgang til BlueIdea",
+      en_US: "You are not allowed to use BlueIdea",
+    },
+    "Starting analysis": {
+      da_DK: "Analyserer",
+      en_US: "Analyzing",
+    },
+    "modify parcels": {
+      da_DK: "Tilføj eller fjern matrikler",
+      en_US: "Add or remove parcels",
+    },
+    "Select profile": {
+      da_DK: "Vælg profil",
+      en_US: "Select profile",
+    },
+    "Alarm cable": {
+      da_DK: "Alarmkabel",
+      en_US: "Alarm cable",
+    },
+    "Distance": {
+      da_DK: "Afstand",
+      en_US: "Distance",
+    },
+    "Distance not set": {
+      da_DK: "Ugyldig afstand",
+      en_US: "Invalid distance",
+    },
+    "Alarm found": {
+      da_DK: "Mulige placeringer fundet",
+      en_US: "Possible alarms found",
+    },
+    "Lukkeliste is ready": {
+      da_DK: "Lukkeliste klar",
+      en_US: "Valve list ready",
+    },
+    "Lukkeliste not ready": {
+      da_DK: "Lukkeliste ikke klar",
+      en_US: "Valve list not ready",
+    },
+    "Large number of parcels found": {
+      da_DK: "Der er fundet for mange matrikler. Matrikler vil ikke blive vist. Der kan gå tid før vi er klar til at hente adresser",
+      en_US: "Too many parcels found. Parcels will not be shown. It may take some time before we are ready to fetch addresses",
+    },
+  };
+
+  module.exports = dict;
